requirements.txt
click>=8.0
rich>=13.0
curl_cffi>=0.7
thefuzz>=0.22
seleniumbase>=4.0
google-genai>=1.0
groq>=0.9
